tests/requirements-test.txt
pytest
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio
httpx>=0.24.1
requests-mock>=1.11.0
faker>=19.3.1
factory-boy>=3.3.0
freezegun>=1.2.2
coverage>=7.3.0
locust>=2.15.1
hypothesis>=6.82.1
testcontainers>=3.7.1
responses>=0.23.1
websockets>=11.0.3
pytest-benchmark>=4.0.0
quart-schema
quart-redis
quart-jwt-extended
fakeredis
redis
uvloop; sys_platform != 'win32'
python-dotenv
google-cloud-video-live-stream>=1.3.1
google-cloud-storage
pytest-env
surrealdb
flake8
novu-py
Pillow
purreal
aiocache
prometheus-client
faststream[rabbit]==0.5.39
stripe
ipinfo